feat: 다른 유저의 댓글이 달린 질문 삭제 불가 기능 추가
- /resources/query.yml 및 Query 열거형에 COUNT_REPLY_BY_ARTICLE_AND_NOT_USER 추가
- ArticleService 내 isDeletable 메서드 추가
- ReplyRepository 내 countByArticleIdAndNotUserId 메서드 추가 및 ReplyJdbcRepository 에서 구현
- ErrorCode 내 INVALID_REPLY_WRITER 추가
- ReplyJdbcRepositoryTest, ReplyServiceTest, ArticleControllerTest 단위/통합 테스트 진행 및 유저가 삭제하지 못하는 테스트 케이스 추가
- SELECT_ARTICLES 에서 댓글이 모두 삭제되었을 때 질문이 사라지는 문제를 SQL 에서 수정하여 해결

+++ src/main/java/com/kakao/cafe/service/ArticleService.java
@@ -86,7 +86,11 @@ public class ArticleService {
     public void deleteArticle(SessionUser user, Integer articleId) {
         findUserArticle(user, articleId);
 
+        if (!isDeletable(user.getUserId(), articleId)) {
+            throw new InvalidRequestException(ErrorCode.INVALID_ARTICLE_DELETE);
+        }
         articleRepository.deleteById(articleId);
+
     }
 
     private Article findUserArticle(SessionUser user, Integer articleId) {
@@ -103,6 +107,10 @@ public class ArticleService {
         if (!article.checkWriter(user.getUserId())) {
             throw new InvalidRequestException(ErrorCode.INVALID_ARTICLE_WRITER);
         }
+    }
+
+    private boolean isDeletable(String userId, Integer articleId) {
+        return replyRepository.countByArticleIdAndNotUserId(userId, articleId) == 0;
     }
 
 }
